Fix for typecast issue causing compile failure with some compilers. Ref issue #761.

+++ report.c
@@ -648,7 +648,7 @@ void report_ngc_parameters (void)
     tool_data_t *tool_data;
 
     for (idx = 1; idx <= grbl.tool_table.n_tools; idx++) {
-        if((tool_data = grbl.tool_table.get_tool_by_idx(idx)) &&
+        if((tool_data = grbl.tool_table.get_tool_by_idx((uint32_t)idx)) &&
             (settings.macro_atc_flags.random_toolchanger ? tool_data->tool_id >= 0 : tool_data->tool_id > 0)) {
             hal.stream.write("[T:");
             hal.stream.write(uitoa(tool_data->tool_id));
